code review suggestions and cleanup
