+++ packages/backend/package.json
@@ -6,7 +6,6 @@
   },
   "dependencies": {
     "@types/cors": "^2.8.13",
-    "bcrypt": "^5.1.0",
     "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
@@ -19,12 +18,12 @@
     "passport-jwt": "^4.0.1",
     "pg": "^8.9.0",
     "uuid": "^9.0.0",
-    "validator": "^13.7.0",
     "winston": "^3.8.2",
     "winston-daily-rotate-file": "^4.7.1",
     "zod": "^3.20.2"
   },
   "devDependencies": {
+    "@faker-js/faker": "^7.6.0",
     "@types/bcryptjs": "^2.4.2",
     "@types/cookie-parser": "^1.4.3",
     "@types/express": "^4.17.17",
