+++ packages/backend/src/index.ts
@@ -1,34 +1,5 @@
-import * as dotenv from 'dotenv'
-
-import { resolve } from 'path'
-import { object, string } from 'zod'
-
-const envPath = resolve(__dirname, '../.env')
-
-const envVars = dotenv.config({ path: envPath })
-
-const envSchema = object({
-  PORT: string(),
-  DB_NAME: string(),
-  DB_USERNAME: string(),
-  DB_PASSWORD: string(),
-  JWT_ACCESS_TOKEN_SECRET: string(),
-  JWT_ACCESS_TOKEN_EXPIRATION_TIME: string(),
-  JWT_REFRESH_TOKEN_SECRET: string(),
-  JWT_REFRESH_TOKEN_EXPIRATION_TIME: string()
-})
-
-if (!envVars || !envVars.parsed) {
-  throw new Error('Could not parse environment variables')
-}
-
-try {
-  envSchema.parse(envVars.parsed)
-} catch (err) {
-  console.error('Error parsing environment variables:', err)
-  process.exit(1)
-}
-
+import { parseEnv } from './config/env.js'
+parseEnv()
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const config = require('../knexfile.js')
 
@@ -39,7 +10,8 @@ import Knex from 'knex'
 import { Model } from 'objection'
 import passport from 'passport'
 import { jwtStrategy } from './auth/jwtStrategy.js'
-import { errorHandler } from './errorHandler'
+
+import { errorHandler } from './middlewares/errorHandler'
 import { mainRouter } from './routes.js'
 
 const app: Application = express()
